# backend/src/learning/dto/learning-lesson-detail-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { LessonDetailDto } from './lesson-detail.dto';

export class LearningLessonDetailResponseDto {
  @ApiProperty({ type: LessonDetailDto })
  lesson: LessonDetailDto;
}