# backend/src/profile/dto/update-profile-request.dto.ts
import { ApiPropertyOptional } from '@nestjs/swagger';
import { IsOptional, IsString } from 'class-validator';

export class UpdateProfileRequestDto {
  @ApiPropertyOptional({ example: 'Thomas' })
  @IsOptional()
  @IsString()
  firstName?: string;

  @ApiPropertyOptional({ example: 'Anderson' })
  @IsOptional()
  @IsString()
  lastName?: string;
}