# docs/02-architecture/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Username or email already taken"
          }
        },
        "summary": "Register a new account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Authenticate user with username and password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "summary": "Refresh access and refresh tokens",
        "tags": [
          "Auth"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The Health Check is successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {},
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "The Health Check is not successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        },
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Service health check",
        "tags": [
          "Health"
        ]
      }
    },
    "/profile/me": {
      "get": {
        "operationId": "ProfileController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get own profile (full)",
        "tags": [
          "Profile"
        ]
      },
      "patch": {
        "operationId": "ProfileController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update own profile (firstName, lastName)",
        "tags": [
          "Profile"
        ]
      }
    },
    "/profile/{id}": {
      "get": {
        "operationId": "ProfileController_getPublicProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProfileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Get public profile by id",
        "tags": [
          "Profile"
        ]
      }
    },
    "/profile/me/password": {
      "patch": {
        "operationId": "ProfileController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password changed"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Wrong current password or unauthorized"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Change password",
        "tags": [
          "Profile"
        ]
      }
    },
    "/profile/me/avatar": {
      "post": {
        "operationId": "ProfileController_uploadAvatar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid file type or size"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Upload avatar",
        "tags": [
          "Profile"
        ]
      },
      "delete": {
        "operationId": "ProfileController_removeAvatar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Remove avatar",
        "tags": [
          "Profile"
        ]
      }
    },
    "/learning/overview": {
      "get": {
        "operationId": "LearningController_getOverview",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningOverviewResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get module and lesson overview for the curriculum",
        "tags": [
          "Learning"
        ]
      }
    },
    "/lessons/{id}": {
      "get": {
        "operationId": "LearningController_getLessonById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningLessonDetailResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token"
          },
          "404": {
            "description": "Lesson was not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get lesson detail content by lesson identifier",
        "tags": [
          "Learning"
        ]
      }
    }
  },
  "info": {
    "title": "Typocalypse Backend API",
    "description": "REST API for Typocalypse terminal training app",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Provide access token as: Bearer <token>"
      }
    },
    "schemas": {
      "RegisterRequestDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "neo"
          },
          "email": {
            "type": "string",
            "example": "neo@zion.dev"
          },
          "password": {
            "type": "string",
            "example": "supersecure"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "7e393f8e-d5f1-4d47-84d4-a1a8759eb0df"
          },
          "username": {
            "type": "string",
            "example": "neo"
          },
          "firstName": {
            "type": "string",
            "example": ""
          },
          "lastName": {
            "type": "string",
            "example": ""
          },
          "email": {
            "type": "string",
            "example": "neo@zion.dev"
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true,
            "example": null
          },
          "createdAt": {
            "type": "string",
            "example": "2026-03-04T10:35:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-03-04T10:35:00.000Z"
          }
        },
        "required": [
          "id",
          "username",
          "firstName",
          "lastName",
          "email",
          "avatarUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "AuthTokensResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.access.payload.signature"
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh.payload.signature"
          },
          "expiresAt": {
            "type": "string",
            "example": "2026-03-04T10:50:00.000Z"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresAt"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "tokens": {
            "$ref": "#/components/schemas/AuthTokensResponseDto"
          }
        },
        "required": [
          "user",
          "tokens"
        ]
      },
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "neo"
          },
          "password": {
            "type": "string",
            "example": "supersecure"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "RefreshRequestDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh.payload.signature"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "UpdateProfileRequestDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Thomas"
          },
          "lastName": {
            "type": "string",
            "example": "Anderson"
          }
        }
      },
      "PublicProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "7e393f8e-d5f1-4d47-84d4-a1a8759eb0df"
          },
          "username": {
            "type": "string",
            "example": "neo"
          },
          "firstName": {
            "type": "string",
            "example": "Thomas"
          },
          "lastName": {
            "type": "string",
            "example": "Anderson"
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true,
            "example": null
          }
        },
        "required": [
          "id",
          "username",
          "firstName",
          "lastName",
          "avatarUrl"
        ]
      },
      "ChangePasswordRequestDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "oldPassword123"
          },
          "newPassword": {
            "type": "string",
            "example": "newPassword456"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "LessonOverviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ls-home"
          },
          "slug": {
            "type": "string",
            "example": "ls-home"
          },
          "title": {
            "type": "string",
            "example": "List home directory"
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "locked",
              "active",
              "completed"
            ],
            "example": "active"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "order",
          "status"
        ]
      },
      "ModuleOverviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmd-basics"
          },
          "slug": {
            "type": "string",
            "example": "cmd-basics"
          },
          "title": {
            "type": "string",
            "example": "Command Line Basics"
          },
          "description": {
            "type": "string",
            "example": "Navigate directories, list contents, and read files with core shell commands."
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "lessons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LessonOverviewDto"
            }
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "description",
          "order",
          "lessons"
        ]
      },
      "LearningOverviewResponseDto": {
        "type": "object",
        "properties": {
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleOverviewDto"
            }
          }
        },
        "required": [
          "modules"
        ]
      },
      "LessonDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ls-home"
          },
          "moduleId": {
            "type": "string",
            "example": "cmd-basics"
          },
          "slug": {
            "type": "string",
            "example": "ls-home"
          },
          "title": {
            "type": "string",
            "example": "List home directory"
          },
          "order": {
            "type": "number",
            "example": 1
          },
          "theoryMarkdown": {
            "type": "string",
            "example": "The `ls` command prints files and folders in the current directory. By default you start in the home directory."
          },
          "taskDescription": {
            "type": "string",
            "example": "Print the list of files in your home directory."
          },
          "hints": {
            "example": [
              "Try `ls` first if you are not sure the file exists."
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "moduleId",
          "slug",
          "title",
          "order",
          "theoryMarkdown",
          "taskDescription"
        ]
      },
      "LearningLessonDetailResponseDto": {
        "type": "object",
        "properties": {
          "lesson": {
            "$ref": "#/components/schemas/LessonDetailDto"
          }
        },
        "required": [
          "lesson"
        ]
      }
    }
  }
}